[REFACTOR] move attribution inputs into own directory
