+++ frontend/src/components/Attribution/Input/AttributionInput.vue
@@ -6,7 +6,7 @@
       :message="t('attributions.add.sameAgainWarning')"
     />
 
-    <AttributionAddFormInputRating
+    <AttributionInputRating
       v-if="attribute.data_type === 'RATING'"
       :model-value="modelValue?.integer_value ?? null"
       :validation="attribute.validation_rule"
@@ -15,7 +15,7 @@
         (val: number | null) => updateModelValue({ integer_value: val })
       "
     />
-    <AttributionAddFormInputNumber
+    <AttributionInputNumber
       v-else-if="attribute.data_type === 'INTEGER'"
       ref="integerInputRef"
       :model-value="modelValue?.integer_value ?? null"
@@ -24,7 +24,7 @@
         (val: number | null) => updateModelValue({ integer_value: val })
       "
     />
-    <AttributionAddFormInputNumber
+    <AttributionInputNumber
       v-else-if="attribute.data_type === 'FLOAT'"
       ref="floatInputRef"
       :model-value="modelValue?.float_value ?? null"
@@ -33,7 +33,7 @@
         (val: number | null) => updateModelValue({ float_value: val })
       "
     />
-    <AttributionAddFormInputText
+    <AttributionInputText
       v-else-if="attribute.data_type === 'TEXT'"
       ref="textInputRef"
       :model-value="modelValue?.text_value ?? null"
@@ -48,21 +48,21 @@
         (val: string | null) => updateModelValue({ text_value: val })
       "
     />
-    <AttributionAddFormInputDate
+    <AttributionInputDate
       v-else-if="attribute.data_type === 'DATE'"
       :model-value="modelValue?.date_value ?? null"
       @update:model-value="
         (val: string | null) => updateModelValue({ date_value: val })
       "
     />
-    <AttributionAddFormInputBoolean
+    <AttributionInputBoolean
       v-else-if="attribute.data_type === 'BOOLEAN'"
       :model-value="modelValue?.boolean_value ?? null"
       @update:model-value="
         (val: boolean | null) => updateModelValue({ boolean_value: val })
       "
     />
-    <AttributionAddFormInputPhoto
+    <AttributionInputPhoto
       v-else-if="attribute.data_type === 'PHOTO'"
       :model-value="modelValue?.photo_value ?? null"
       @update:model-value="
@@ -75,7 +75,7 @@
     </div>
 
     <template v-if="!hideNotes">
-      <AttributionAddFormInputNote
+      <AttributionInputNote
         ref="noteInputRef"
         v-model:photo-note="photoNote"
         v-model:text-note="textNote"
@@ -122,25 +122,25 @@ import type { AttributionValueWithPhoto } from 'src/components/Attribution/Add/A
 import type { AttributeFragment } from 'src/components/Attribute/attributeFragment';
 import BaseInputLabel from 'src/components/Base/BaseInputLabel.vue';
 import BaseMessage from 'src/components/Base/BaseMessage.vue';
-import AttributionAddFormInputRating from 'src/components/Attribution/Add/AttributionAddFormInputRating.vue';
-import AttributionAddFormInputNumber from 'src/components/Attribution/Add/AttributionAddFormInputNumber.vue';
-import AttributionAddFormInputText from 'src/components/Attribution/Add/AttributionAddFormInputText.vue';
-import AttributionAddFormInputDate from 'src/components/Attribution/Add/AttributionAddFormInputDate.vue';
-import AttributionAddFormInputBoolean from 'src/components/Attribution/Add/AttributionAddFormInputBoolean.vue';
-import AttributionAddFormInputPhoto from 'src/components/Attribution/Add/AttributionAddFormInputPhoto.vue';
-import AttributionAddFormInputNote from 'src/components/Attribution/Add/AttributionAddFormInputNote.vue';
+import AttributionInputRating from 'src/components/Attribution/Input/AttributionInputRating.vue';
+import AttributionInputNumber from 'src/components/Attribution/Input/AttributionInputNumber.vue';
+import AttributionInputText from 'src/components/Attribution/Input/AttributionInputText.vue';
+import AttributionInputDate from 'src/components/Attribution/Input/AttributionInputDate.vue';
+import AttributionInputBoolean from 'src/components/Attribution/Input/AttributionInputBoolean.vue';
+import AttributionInputPhoto from 'src/components/Attribution/Input/AttributionInputPhoto.vue';
+import AttributionInputNote from 'src/components/Attribution/Input/AttributionInputNote.vue';
 import { computed, ref, nextTick } from 'vue';
 import { useI18n } from 'src/composables/useI18n';
 import type { DistributiveOmit } from 'src/utils/typescriptUtils';
 
-export interface AttributionAddFormInputProps {
+export interface AttributionInputProps {
   attribute: DistributiveOmit<AttributeFragment, 'created' | 'modified'>;
   exceptional: boolean;
   hasSameAgain: boolean;
   hideNotes?: boolean;
 }
 
-const props = defineProps<AttributionAddFormInputProps>();
+const props = defineProps<AttributionInputProps>();
 
 const modelValue = defineModel<AttributionValueWithPhoto | undefined>({
   required: true,
@@ -160,9 +160,9 @@ const textNote = computed({
 
 const { t } = useI18n();
 const confirm = ref(false);
-const noteInputRef = ref<InstanceType<
-  typeof AttributionAddFormInputNote
-> | null>(null);
+const noteInputRef = ref<InstanceType<typeof AttributionInputNote> | null>(
+  null,
+);
 
 async function clearModelValue() {
   if (!modelValue.value) {
@@ -259,15 +259,15 @@ const hasNoValue = computed(() => {
   );
 });
 
-const integerInputRef = ref<InstanceType<
-  typeof AttributionAddFormInputNumber
-> | null>(null);
-const floatInputRef = ref<InstanceType<
-  typeof AttributionAddFormInputNumber
-> | null>(null);
-const textInputRef = ref<InstanceType<
-  typeof AttributionAddFormInputText
-> | null>(null);
+const integerInputRef = ref<InstanceType<typeof AttributionInputNumber> | null>(
+  null,
+);
+const floatInputRef = ref<InstanceType<typeof AttributionInputNumber> | null>(
+  null,
+);
+const textInputRef = ref<InstanceType<typeof AttributionInputText> | null>(
+  null,
+);
 
 async function validate() {
   let valueIsValid: Promise<boolean> = Promise.resolve(true);
