+++ frontend/src/components/Attribution/Input/AttributionInputBoolean.vue
