+++ frontend/src/components/Attribution/Input/AttributionInputPhoto.vue
